fix support for projects with different `python'/'labextension_name`
I just ran into another breaking bug in the new hatch toml. This is hidden if the names of the project's python and js packages are exactly the same. However, when I made a project with python package `jupyterlab_x_y` and js package `jupyterlab-x-y` I ran into pathing hell. Ultimately the issue in this case is that the javascript machinery that loads the extension bundle files expect all of the files to be under a dir named after the js package (eg `share/jupyter/labextensions/jupyterlab-x-y`), but the toml template instead places the bundle files under a dir named after the python package (eg `share/jupyter/labextensions/jupyterlab_x_y`). This leads to a fatal NOT-FOUND error on attempted extension bundle load. This PR fixes that